Remove clipper-base class
